Refactoring filepath
